app/requirements.txt
pyyaml
requests
flask
